+++ tests/test_dipdup/test_models.py
@@ -18,6 +18,8 @@ from tests.test_dipdup.types.bazaar.storage import BazaarMarketPlaceStorage
 from tests.test_dipdup.types.ftzfun.storage import FtzFunStorage
 from tests.test_dipdup.types.hen_subjkt.storage import HenSubjktStorage
 from tests.test_dipdup.types.hjkl.storage import HjklStorage
+from tests.test_dipdup.types.kolibri_ovens.set_delegate import SetDelegateParameter
+from tests.test_dipdup.types.kolibri_ovens.storage import KolibriOvensStorage
 from tests.test_dipdup.types.listofmaps.storage import ListOfMapsStorage
 from tests.test_dipdup.types.qwer.storage import QwerStorage
 from tests.test_dipdup.types.rewq.storage import RewqStorage
@@ -334,3 +336,17 @@ class ModelsTest(TestCase):
         self.assertIsInstance(storage_obj, HenSubjktStorage)
         self.assertIsInstance(storage_obj.entries, dict)
         self.assertEqual(storage_obj.entries['tz1Y1j7FK1X9Rrv2VdPz5bXoU7SszF8W1RnK'], True)  # type: ignore
+
+    def test_kolibri_ovens(self) -> None:
+        with open(join(dirname(__file__), 'kolibri_ovens.json')) as f:
+            operations_json = json.load(f)
+
+        # Act
+        operations = [TzktDatasource.convert_operation(op) for op in operations_json]
+        storage_obj = deserialize_storage(operations[0], KolibriOvensStorage)
+        parameter_obj = SetDelegateParameter.parse_obj(operations[0].parameter_json)
+
+        # Assert
+        self.assertIsInstance(storage_obj, KolibriOvensStorage)
+        self.assertIsInstance(parameter_obj, SetDelegateParameter)
+        self.assertEqual(parameter_obj.__root__, None)